+++ frontend/src/metabase-lib/lib/ValidationError/constants.ts
@@ -1,0 +1,3 @@
+export const VALIDATION_ERROR_TYPES = {
+  MISSING_TAG_DIMENSION: "MISSING_TAG_DIMENSION",
+} as const;
