+++ frontend/src/metabase-lib/lib/ValidationError/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./ValidationError";
+export * from "./constants";
+export * from "./types";
