+++ frontend/src/metabase-lib/lib/ValidationError/types.ts
@@ -1,0 +1,3 @@
+import { VALIDATION_ERROR_TYPES } from "./constants";
+
+export type ErrorType = typeof VALIDATION_ERROR_TYPES[keyof typeof VALIDATION_ERROR_TYPES];
